Reapply "[SandboxIR][SandboxVec] Remove score tracking from Region, add RegionWithScore" (#195708)
This reverts commit 75f884b96dfa082227fb730435c68353cd3376c5.
