+++ llvm/include/llvm/SandboxIR/Region.h
@@ -14,47 +14,12 @@
 
 #include "llvm/ADT/SetVector.h"
 #include "llvm/ADT/iterator_range.h"
-#include "llvm/Analysis/TargetTransformInfo.h"
+#include "llvm/SandboxIR/BasicBlock.h"
+#include "llvm/SandboxIR/Function.h"
 #include "llvm/SandboxIR/Instruction.h"
 #include "llvm/Support/raw_ostream.h"
 
 namespace llvm::sandboxir {
-
-class Region;
-
-class ScoreBoard {
-  const Region &Rgn;
-  TargetTransformInfo &TTI;
-  constexpr static TTI::TargetCostKind CostKind = TTI::TCK_RecipThroughput;
-  /// The cost of all instructions added to the region.
-  InstructionCost AfterCost = 0;
-  /// The cost of all instructions that got removed and replaced by new ones.
-  InstructionCost BeforeCost = 0;
-  /// Helper for both add() and remove(). \Returns the TTI cost of \p I.
-  LLVM_ABI InstructionCost getCost(Instruction *I) const;
-  /// No need to allow copies.
-  ScoreBoard(const ScoreBoard &) = delete;
-  const ScoreBoard &operator=(const ScoreBoard &) = delete;
-
-public:
-  ScoreBoard(Region &Rgn, TargetTransformInfo &TTI) : Rgn(Rgn), TTI(TTI) {}
-  /// Mark \p I as a newly added instruction to the region.
-  void add(Instruction *I) { AfterCost += getCost(I); }
-  /// Mark \p I as a deleted instruction from the region.
-  LLVM_ABI void remove(Instruction *I);
-  /// \Returns the cost of the newly added instructions.
-  InstructionCost getAfterCost() const { return AfterCost; }
-  /// \Returns the cost of the Removed instructions.
-  InstructionCost getBeforeCost() const { return BeforeCost; }
-
-#ifndef NDEBUG
-  void dump(raw_ostream &OS) const {
-    OS << "BeforeCost: " << BeforeCost << "\n";
-    OS << "AfterCost:  " << AfterCost << "\n";
-  }
-  LLVM_DUMP_METHOD void dump() const;
-#endif // NDEBUG
-};
 
 /// The main job of the Region is to point to new instructions generated by
 /// vectorization passes. It is the unit that RegionPasses operate on with their
@@ -93,7 +58,13 @@ public:
 // This can be used to pass auxiliary information across region passes, like for
 // example the initial seed slice used by the bottom-up vectorizer.
 
+enum RegionClassID {
+  RegionID,
+  RegionWithScoreID,
+};
+
 class Region {
+protected:
   /// All the instructions in the Region. Only new instructions generated during
   /// vectorization are part of the Region.
   SetVector<Instruction *> Insts;
@@ -107,31 +78,34 @@ class Region {
   static constexpr const char *AuxMDKind = "sandboxaux";
 
   Context &Ctx;
-  /// Keeps track of cost of instructions added and removed.
-  ScoreBoard Scoreboard;
+
+  RegionClassID ID;
 
   /// ID (for later deregistration) of the "create instruction" callback.
   Context::CallbackID CreateInstCB;
   /// ID (for later deregistration) of the "erase instruction" callback.
   Context::CallbackID EraseInstCB;
 
-  /// Adds \p I to the set but also don't track the instruction's score if \p
-  /// IgnoreCost is true. Only to be used when adding an instruction to the
-  /// auxiliary vector.
+  /// Adds \p I to the set. Only to be used when we want to avoid the additional
+  /// functionality provided by the subclass, e.g., to avoid score counting when
+  /// adding instrs to the aux vector.
   /// NOTE: When an instruction is added to the region we track it cost in the
   /// scoreboard, which currently resides in the region class. However, when we
   /// add an instruction to the auxiliary vector it does get tagged as being a
   /// member of the region (for ownership reasons), but its cost does not get
   /// counted because the instruction hasn't been added in the "normal" way.
-  LLVM_ABI void addImpl(Instruction *I, bool IgnoreCost);
+  void addRaw(Instruction *I) {
+    Insts.insert(I);
+    // TODO: Consider tagging instructions lazily.
+    cast<llvm::Instruction>(I->Val)->setMetadata(MDKind, RegionMDN);
+  }
   /// Adds I to the set. This is the main API for adding an instruction to the
   /// region.
-  void add(Instruction *I) { addImpl(I, /*IgnoreCost=*/false); }
+  virtual void add(Instruction *I) { addRaw(I); }
   /// Removes I from the set.
-  LLVM_ABI void remove(Instruction *I);
+  LLVM_ABI virtual void remove(Instruction *I);
   friend class Context; // The callbacks need to call add() and remove().
   friend class RegionInternalsAttorney; // For unit tests.
-  friend class RegionsFromBBs;          // For add().
 
   /// Set \p I as the \p Idx'th element in the auxiliary vector.
   /// NOTE: This is for internal use, it does not set the metadata.
@@ -141,9 +115,56 @@ class Region {
   /// Remove instruction \p I from Aux and drop metadata.
   void removeFromAux(Instruction *I);
 
+  LLVM_ABI Region(Context &Ctx, RegionClassID ID);
+
+  template <typename RegionT, typename RegionFactoryT>
+  static SmallVector<std::unique_ptr<RegionT>>
+  createRegionsFromMD(Function &F, RegionFactoryT Factory) {
+    SmallVector<std::unique_ptr<RegionT>> Regions;
+    DenseMap<MDNode *, RegionT *> MDNToRegion;
+    for (BasicBlock &BB : F) {
+      for (Instruction &Inst : BB) {
+        auto *LLVMI = cast<llvm::Instruction>(Inst.Val);
+        RegionT *R = nullptr;
+        if (auto *MDN = LLVMI->getMetadata(MDKind)) {
+          auto [It, Inserted] = MDNToRegion.try_emplace(MDN);
+          if (Inserted) {
+            Regions.push_back(Factory());
+            R = Regions.back().get();
+            It->second = R;
+          } else {
+            R = It->second;
+          }
+          R->addRaw(&Inst);
+        }
+        if (auto *AuxMDN = LLVMI->getMetadata(AuxMDKind)) {
+          llvm::Constant *IdxC =
+              dyn_cast<ConstantAsMetadata>(AuxMDN->getOperand(0))->getValue();
+          auto Idx = cast<llvm::ConstantInt>(IdxC)->getSExtValue();
+          if (R == nullptr) {
+            errs() << "No region specified for Aux: '" << *LLVMI << "'\n";
+            reportFatalUsageError("No region specified for Aux!");
+          }
+          R->setAux(Idx, &Inst);
+        }
+      }
+    }
+#ifndef NDEBUG
+    // Check that there are no gaps in the Aux vector.
+    for (auto &RPtr : Regions)
+      for (auto *I : RPtr->getAux())
+        assert(I != nullptr && "Gap in Aux!");
+#endif
+    return Regions;
+  }
+
 public:
-  LLVM_ABI Region(Context &Ctx, TargetTransformInfo &TTI);
-  LLVM_ABI ~Region();
+  LLVM_ABI Region(Context &Ctx) : Region(Ctx, RegionClassID::RegionID) {}
+  LLVM_ABI virtual ~Region();
+  // Disable copies to avoid unregistering the callback more than once.
+  Region(const Region &) = delete;
+  // Moves are allowed.
+  Region(Region &&) = default;
 
   Context &getContext() const { return Ctx; }
   /// Returns true if I is in the Region.
@@ -163,9 +184,9 @@ public:
   iterator_range<iterator> insts() { return make_range(begin(), end()); }
 
   LLVM_ABI static SmallVector<std::unique_ptr<Region>>
-  createRegionsFromMD(Function &F, TargetTransformInfo &TTI);
-  /// \Returns the ScoreBoard data structure that keeps track of instr costs.
-  const ScoreBoard &getScoreboard() const { return Scoreboard; }
+  createRegionsFromMD(Function &F);
+
+  RegionClassID getSubclassID() const { return ID; }
 
 #ifndef NDEBUG
   /// This is an expensive check, meant for testing.
