+++ llvm/lib/SandboxIR/Region.cpp
@@ -7,39 +7,16 @@
 //===----------------------------------------------------------------------===//
 
 #include "llvm/SandboxIR/Region.h"
-#include "llvm/SandboxIR/Function.h"
 
 namespace llvm::sandboxir {
 
-InstructionCost ScoreBoard::getCost(Instruction *I) const {
-  auto *LLVMI = cast<llvm::Instruction>(I->Val);
-  SmallVector<const llvm::Value *> Operands(LLVMI->operands());
-  return TTI.getInstructionCost(LLVMI, Operands, CostKind);
-}
-
-void ScoreBoard::remove(Instruction *I) {
-  auto Cost = getCost(I);
-  if (Rgn.contains(I))
-    // If `I` is one the newly added ones, then we should adjust `AfterCost`
-    AfterCost -= Cost;
-  else
-    // If `I` is one of the original instructions (outside the region) then it
-    // is part of the original code, so adjust `BeforeCost`.
-    BeforeCost += Cost;
-}
-
-#ifndef NDEBUG
-void ScoreBoard::dump() const { dump(dbgs()); }
-#endif
-
-Region::Region(Context &Ctx, TargetTransformInfo &TTI)
-    : Ctx(Ctx), Scoreboard(*this, TTI) {
+Region::Region(Context &Ctx, RegionClassID ID) : Ctx(Ctx), ID(ID) {
   LLVMContext &LLVMCtx = Ctx.LLVMCtx;
   auto *RegionStrMD = MDString::get(LLVMCtx, RegionStr);
   RegionMDN = MDNode::getDistinct(LLVMCtx, {RegionStrMD});
 
   CreateInstCB = Ctx.registerCreateInstrCallback(
-      [this](Instruction *NewInst) { add(NewInst); });
+      [this](Instruction *NewInst) { addRaw(NewInst); });
   EraseInstCB = Ctx.registerEraseInstrCallback([this](Instruction *ErasedInst) {
     remove(ErasedInst);
     removeFromAux(ErasedInst);
@@ -49,15 +26,6 @@ Region::Region(Context &Ctx, TargetTransformInfo &TTI)
 Region::~Region() {
   Ctx.unregisterCreateInstrCallback(CreateInstCB);
   Ctx.unregisterEraseInstrCallback(EraseInstCB);
-}
-
-void Region::addImpl(Instruction *I, bool IgnoreCost) {
-  Insts.insert(I);
-  // TODO: Consider tagging instructions lazily.
-  cast<llvm::Instruction>(I->Val)->setMetadata(MDKind, RegionMDN);
-  if (!IgnoreCost)
-    // Keep track of the instruction cost.
-    Scoreboard.add(I);
 }
 
 void Region::setAux(ArrayRef<Instruction *> Aux) {
@@ -71,7 +39,7 @@ void Region::setAux(ArrayRef<Instruction *> Aux) {
     cast<llvm::Instruction>(I->Val)->setMetadata(
         AuxMDKind, MDNode::get(LLVMCtx, ConstantAsMetadata::get(IdxC)));
     // Aux instrs should always be in a region.
-    addImpl(I, /*DontTrackCost=*/true);
+    addRaw(I);
   }
 }
 
@@ -88,7 +56,7 @@ void Region::setAux(unsigned Idx, Instruction *I) {
   }
   Aux[Idx] = I;
   // Aux instrs should always be in a region.
-  addImpl(I, /*DontTrackCost=*/true);
+  addRaw(I);
 }
 
 void Region::dropAuxMetadata(Instruction *I) {
@@ -111,10 +79,6 @@ void Region::clearAux() {
 }
 
 void Region::remove(Instruction *I) {
-  // Keep track of the instruction cost. This need to be done *before* we remove
-  // `I` from the region.
-  Scoreboard.remove(I);
-
   Insts.remove(I);
   cast<llvm::Instruction>(I->Val)->setMetadata(MDKind, nullptr);
 }
@@ -148,45 +112,9 @@ void Region::dump() const {
 }
 #endif // NDEBUG
 
-SmallVector<std::unique_ptr<Region>>
-Region::createRegionsFromMD(Function &F, TargetTransformInfo &TTI) {
-  SmallVector<std::unique_ptr<Region>> Regions;
-  DenseMap<MDNode *, Region *> MDNToRegion;
-  auto &Ctx = F.getContext();
-  for (BasicBlock &BB : F) {
-    for (Instruction &Inst : BB) {
-      auto *LLVMI = cast<llvm::Instruction>(Inst.Val);
-      Region *R = nullptr;
-      if (auto *MDN = LLVMI->getMetadata(MDKind)) {
-        auto [It, Inserted] = MDNToRegion.try_emplace(MDN);
-        if (Inserted) {
-          Regions.push_back(std::make_unique<Region>(Ctx, TTI));
-          R = Regions.back().get();
-          It->second = R;
-        } else {
-          R = It->second;
-        }
-        R->addImpl(&Inst, /*IgnoreCost=*/true);
-      }
-      if (auto *AuxMDN = LLVMI->getMetadata(AuxMDKind)) {
-        llvm::Constant *IdxC =
-            dyn_cast<ConstantAsMetadata>(AuxMDN->getOperand(0))->getValue();
-        auto Idx = cast<llvm::ConstantInt>(IdxC)->getSExtValue();
-        if (R == nullptr) {
-          errs() << "No region specified for Aux: '" << *LLVMI << "'\n";
-          reportFatalUsageError("No region specified for Aux!");
-        }
-        R->setAux(Idx, &Inst);
-      }
-    }
-  }
-#ifndef NDEBUG
-  // Check that there are no gaps in the Aux vector.
-  for (auto &RPtr : Regions)
-    for (auto *I : RPtr->getAux())
-      assert(I != nullptr && "Gap in Aux!");
-#endif
-  return Regions;
+SmallVector<std::unique_ptr<Region>> Region::createRegionsFromMD(Function &F) {
+  return Region::createRegionsFromMD<Region>(
+      F, [&F]() { return std::make_unique<Region>(F.getContext()); });
 }
 
 } // namespace llvm::sandboxir
