+++ llvm/lib/Transforms/Vectorize/SandboxVectorizer/Passes/TransactionAcceptOrRevert.cpp
@@ -10,6 +10,7 @@
 #include "llvm/Support/CommandLine.h"
 #include "llvm/Support/InstructionCost.h"
 #include "llvm/Transforms/Vectorize/SandboxVectorizer/Debug.h"
+#include "llvm/Transforms/Vectorize/SandboxVectorizer/RegionWithScore.h"
 
 namespace llvm {
 
@@ -20,7 +21,7 @@ static cl::opt<int> CostThreshold("sbvec-cost-threshold", cl::init(0),
 namespace sandboxir {
 
 bool TransactionAcceptOrRevert::runOnRegion(Region &Rgn, const Analyses &A) {
-  const auto &SB = Rgn.getScoreboard();
+  const auto &SB = cast<RegionWithScore>(Rgn).getScoreboard();
   [[maybe_unused]] auto CostBefore = SB.getBeforeCost();
   [[maybe_unused]] auto CostAfter = SB.getAfterCost();
   InstructionCost CostAfterMinusBefore = SB.getAfterCost() - SB.getBeforeCost();
